chore: move UIViewController snapshot helpers to separate file
